refactor(peg): Different approach for matching recursive lists
Counted the depth of list-within-list & used that to `car' the
`current' element of the input. Used that approach for `scanl.next'
and `scanl.currc'.

On the way of the above refactoring, found out that there was no
implementation of the primitive `any' whatsoever, provided it to
`scan` and to `scanl'. The latter one was required for parsing lists
and data structures because how we get the Atom via `!{ .* } .`.

One last thingy on refactoring, the predicate was removed from the
output of `match' & `matchl' by returning an instance of the class
Predicate and filtered out by `matchexpr'.