refactor(peg): yet another change to matching recursive lists
No depth counting. Use a stack to save the tail of the subject and set
the head of the subject as the new subject when open the list.  Then,
after the matching of the elements is done, pop the last stack entry
from the list stack and set it as the new subject.  This technique was
seen in the article Parsing Expression Grammars for Structured Data.

The tests for parsing atoms are still broken. :(