feat(py37): Compile first lil function call to bytecode \o/ \o/ \o/
Yay! `print(42)` finally works. The generated .pyc can be interpreted
by Python3.7 interpreters.

There's a lot of hardcoded stuff though. The hardest one to figure
out will probably be `co_stacksize' and that's needed for whatever
comes next.

The size of the buffer that the final bytecode is written isn't
figured out yet. It's all hardcoded for this one case for now.

Most interesting parts of the patch were on the ripoff of the marshal:

 1. wRef's cache needed to be changed to an object because it can hash
    arrays uniquely and Maps can't. However, objects hashed empty
    strings and empty arrays to the same id, so the gotcha was to
    cache `[typeof v, v]` so we avoid the weird clash.

 2. Not really sure why but wTYPE didn't work for small tuple and had
    to use wByte. My best guest is signedness but I'm really not clear
    on this one.