refactor(peg): Extract action executor from pattern match function
The match to be passed to the action was wrapped in a thunk so a
custom action would be able to delay the execution of the thunk until
it chose to.

Since this feature implies that user will have to care about how
backtrack works, it is made optional and not the default.