feat(lang): Most of scope analysis seems to be in place
Tests are stil breaking but it's unrelated to the scope work. Besides
just fixing the analysis algorithm, the walking of the scope tree in
the code generator traversal was fixed as well and the cell vars were
added to the code object as well as the load-closure instructions for
the deref vars.