feat(lang): Comparison Operators
These aren't like Python's that accept math like expressions with
multiple operators such as `a > b > c`. This simple implementation is
like the C family where `a > b` evaluates to a boolean value and then
true or false won't make sense to be used with `c`. But that will do
for now.