feat(py37): Calculate stack depth for each code object
This is the simplest code to achieve the minimum needed to do this
task. There are some incomplete things that might lead to runtime
errors on scripts that look good. The most pressing example that might
not work is stuff that involve jump instructions that I'm currently
ignoring.

Most of the important parts of this code were copied from the Python
source code (compile.c) and adapted to our weird JavaScript world.